+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/mmogroup/mmoitems/MMOItemsUtils.java
@@ -1,11 +1,15 @@
 package org.betonquest.betonquest.compatibility.mmogroup.mmoitems;
 
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 import io.lumine.mythic.lib.api.item.NBTItem;
 import net.Indyuce.mmoitems.MMOItems;
 import net.Indyuce.mmoitems.api.Type;
 import org.betonquest.betonquest.api.QuestException;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.Nullable;
+
+import java.util.UUID;
 
 /**
  * A utility class for working with MMOItems.
@@ -21,20 +25,33 @@ public final class MMOItemsUtils {
     /**
      * Compares an ItemStack and an MMOItem based on the MMOItem's type and ID.
      *
-     * @param item   any ItemStack.
-     * @param type   the type of the MMOItem that shall be checked for
-     * @param itemID the item of the MMOItem that shall be checked for
+     * @param item       any ItemStack
+     * @param type       the type of the MMOItem that shall be checked for
+     * @param itemID     the item of the MMOItem that shall be checked for
+     * @param createdFor the soulbound uuid to match, if present
      * @return whether the input item matches the defined MMOItems
      */
-    public static boolean equalsMMOItem(@Nullable final ItemStack item, final Type type, final String itemID) {
+    public static boolean equalsMMOItem(@Nullable final ItemStack item, final Type type, final String itemID,
+                                        @Nullable final UUID createdFor) {
         if (item == null) {
             return false;
         }
         final NBTItem realItemNBT = NBTItem.get(item);
         final String realItemType = realItemNBT.getString("MMOITEMS_ITEM_TYPE");
         final String realItemID = realItemNBT.getString("MMOITEMS_ITEM_ID");
-
-        return realItemID.equalsIgnoreCase(itemID) && realItemType.equalsIgnoreCase(type.getId());
+        if (!realItemID.equalsIgnoreCase(itemID) || !realItemType.equalsIgnoreCase(type.getId())) {
+            return false;
+        }
+        if (createdFor == null) {
+            return true;
+        }
+        final String soulbound = realItemNBT.hasTag("MMOITEMS_SOULBOUND") ? realItemNBT.getString("MMOITEMS_SOULBOUND") : null;
+        if (soulbound == null) {
+            return false;
+        }
+        final JsonObject json = JsonParser.parseString(soulbound).getAsJsonObject();
+        final String ownerUuid = json.get("UUID").getAsString();
+        return ownerUuid.equals(createdFor.toString());
     }
 
     /**