+++ CHANGELOG.md
@@ -14,6 +14,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 - `language` config option is now a section; the default language moved to `language.default`
 - `burn` action to no longer require its duration argument to specify its name `duration`
+- `mmoitem` now checks for soulbound when set
 ### Deprecated
 ### Removed
 ### Fixed